% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circRNAFull.R
\name{extract_exon_after_delete_skip_exon}
\alias{extract_exon_after_delete_skip_exon}
\title{Extracting exon after deleting skip exon}
\usage{
extract_exon_after_delete_skip_exon(
  exon_count_file,
  skip_exon_file,
  output_folder
)
}
\arguments{
\item{exon_count_file}{A data frame containing exons for circRNAs (obtained from function \code{\link[circRNAFull]{Extract_exon}})}

\item{skip_exon_file}{A data frame containing skip exons for circRNAs (obtained from function \code{\link[circRNAFull]{skip_exon_detection}})}

\item{output_folder}{The name of the output directory}
}
\value{
A text file containing the exons after deleting skip exon
}
\description{
This function extract the remaining exons after deleting skip exons
}
\examples{
#loading an example exon count file
data(exon_count)
exon_count_file<-exon_count

#Loading an example skip exon file
data(skip_exon)
skip_exon_file<-skip_exon

#Creating a temporary output directory
output_folder<-tempdir()

#Extracting exon after deleting skip exon
extract_exon_after_delete_skip_exon(exon_count_file, skip_exon_file, output_folder)
}
