% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circRNAFull.R
\name{extract_sequence}
\alias{extract_sequence}
\title{Reconstruction of full length circRNA sequences}
\usage{
extract_sequence(seq_name, sequence, exon_count_file, output)
}
\arguments{
\item{seq_name}{A vector containing the name of chromosomes of the reference genome}

\item{sequence}{A vector containing the sequences of the reference genome}

\item{exon_count_file}{A data frame containing exons after deleting skip exons}

\item{output}{The output folder name}
}
\value{
The reconstructed circRNA sequences will be written in folder 'output'
}
\description{
This function can reconstruct the full length circRNA
sequences from the output of the circular RNA predictions tools
}
\examples{
\donttest{
\dontrun{
#Please download the reference genome. suppose you have downloaded
#the reference genome 'hg38.fa' in you working directory.
#Then run:
ref_genome<-"hg38.fa"
fastaFile <- readDNAStringSet(ref_genome)
seq_name = sub('\\\\ .*', '', names(fastaFile))
sequence = paste(fastaFile)
}
}

#Loading an exon count file after deleting skip exon
exon_count<-data(exon_count_after_skipping_exon)
exon_count<-exon_count_after_skipping_exon

#Creating an output directory
output<-tempdir()

\donttest{
\dontrun{
#Reconstructing the circRNA sequences. The circRNA sequences will be generated
#in the 'output' directory.
extract_sequence(seq_name, sequence, exon_count, output)
}
}
}
