% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circRNAFull.R
\name{skip_exon_detection}
\alias{skip_exon_detection}
\title{Detecting skip exon}
\usage{
skip_exon_detection(circle_reads_folder, exon_count_file, output_folder)
}
\arguments{
\item{circle_reads_folder}{Location of the individual alignments (obtained from function \code{\link[circRNAFull]{extract_reads_from_bam}})}

\item{exon_count_file}{A data frame containing exons for circRNAs (obtained from function \code{\link[circRNAFull]{Extract_exon}})}

\item{output_folder}{The name of the output directory}
}
\value{
A file containing the skip exons written in 'output_folder'
}
\description{
This function detects skip exon
}
\examples{
#Example of circle_reads_folder containing the individual alignments
circle_reads_folder<-system.file('extdata', package = 'circRNAFull')

#Loading an example exon_count_file containing exons for circRNAs
data(exon_count)
exon_count_file<-exon_count

#Creating a temporary output directory
output_folder<-tempdir()

#Detecting skip exon
skip_exon_detection(circle_reads_folder, exon_count_file, output_folder)
}
